% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spendingCategories.R
\name{addCategories}
\alias{addCategories}
\title{addCategories}
\usage{
addCategories(Description, myCategories)
}
\arguments{
\item{Description}{character vector}

\item{myCategories}{list; each named element contains a character
vector of all known patterns defining that category.}
}
\value{
character vector
}
\description{
For each budget category check if in any pattern is a
 match with the \code{Description}.
 Return the name of the category that matches.
}
\details{
If none of the categories contain a pattern that matches
 then return \code{MISSING}. A \code{Description} is not allowed to
 belong to multiple categories. The first instance of such
 a problem will cause an error to be thrown with the first
 \code{Description} and all the categories it has been matched
 with.
}
\examples{
  Description <- c('rent 2019 march',
                   'Happy Co. gas 20190601',
                   'New restaurant')
  myCategories <- list(housing = c('rent', 'Hardware Inc'),
                       transportation = 'gas')
  addCategories(Description = Description, myCategories)
\dontrun{
  addCategories(Description = c("gas rent"), myCategories)
}
}
