% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveLoad.R
\name{appendCreditCardData}
\alias{appendCreditCardData}
\title{appendCreditCardData}
\usage{
appendCreditCardData(...)
}
\arguments{
\item{...}{data.table(s) of credit card records}
}
\description{
Combine a sequence of data.tables by row to
  append new credit card data to the previous credit card
  data.
}
\details{
Duplicate records based on matching \code{'PostingDate', 'Amount', 'Description'}
  are removed to avoid double counting.
}
