% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-categoryPatternsExample.R
\docType{data}
\name{categoryPatternsExample}
\alias{categoryPatternsExample}
\title{Home budget sample spending category patterns}
\format{
A list with 9 named character vectors
\describe{
\item{Grocery}{Food shopping}
\item{Housing}{Rent, home repair, home insurance, mortgage installments}
\item{Entertainment}{Discretionary spending (excluding meals)}
\item{Restaurant}{Discretionary spending for meals}
\item{Travel}{Vacation expenses, airfare, hotel}
\item{Utilities}{water, sewer, garbage, gas, electric, cable, internet}
\item{Transportation}{Vehicle maitenence, fuel, automotive insurance,
licensing fees, bus fare}
\item{Donation}{Charitable giving}
\item{Medical}{Medical & Dental health insurance: fixed monthly premiums,
out-of-pocket expenses, emergency visits}
}
}
\usage{
categoryPatternsExample
}
\description{
This data is a sample of categories for breaking
down expenses. Each group contains a string
of patterns matching the description in the credit
card statetement.
}
\examples{
categoryPatternsExample
}
\keyword{datasets}
