% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-creditCardExample.R
\docType{data}
\name{creditCardExample}
\alias{creditCardExample}
\title{Home budget sample credit card data}
\format{A data.table with 46 rows and 3 variables
\describe{
\item{PostingDate}{date the credit card transaction is posted to
the card holder's account, typically the same as the date of the
transaction}
\item{Description}{text describing the company purchased from}
\item{Amount}{dollars posted to card balance}
}}
\usage{
creditCardExample
}
\description{
This data is a sample of purchasing data.
}
\examples{
creditCardExample
}
\keyword{datasets}
