% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homeBudget.R
\docType{package}
\name{homeBudget}
\alias{homeBudget}
\alias{homeBudget-package}
\title{homeBudget: A package for computating the personal finances.}
\description{
The homeBudget package provides three categories of important functions:
foo, bar and baz.
}
\section{homeBudget functions}{

The homeBudget functions ...
}

