% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarterlyIncome.R
\name{length_overlap}
\alias{length_overlap}
\title{length_overlap}
\usage{
length_overlap(interval1, interval2, units)
}
\arguments{
\item{interval1, interval2}{lubridate interval object with
specified start and end dates.}

\item{units}{character string. Units in which the results are desired.}
}
\description{
Count the amount of overlap between two date intervals.
}
\examples{
int1 <- lubridate::interval(start = as.Date(c("2019-01-01", "2019-02-01")),
                            end = as.Date(c("2019-03-01", "2019-05-01")))
int2 <- lubridate::interval(start = as.Date("2019-03-02"),
                            end = as.Date("2019-04-02"))
length_overlap(interval1 = int1, interval2 = int2, units = "days")

}
