% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveLoad.R
\name{loadCreditCard}
\alias{loadCreditCard}
\title{loadCreditCard}
\usage{
loadCreditCard(path)
}
\arguments{
\item{path}{the path to the folder for the saved R credit card}
}
\value{
data.table of saved credit card history
}
\description{
Load the credit card history data \code{creditCardHistory}
  saved in the R object \code{creditCardHistory.rda}.
}
\examples{
  \dontrun{
    loadCreditCard(path = "/budget/data")
   }
}
