% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveLoad.R
\name{modifyCreditCard_AmEx}
\alias{modifyCreditCard_AmEx}
\title{modifyCreditCard_AmEx}
\usage{
modifyCreditCard_AmEx(dataCredit)
}
\description{
Modify American Express to exclude monthly payments.
}
\keyword{internal}
