% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveLoad.R
\name{modifyCreditCard_Chase}
\alias{modifyCreditCard_Chase}
\title{modifyCreditCard_Chase}
\usage{
modifyCreditCard_Chase(dataCredit)
}
\description{
Modify Chase to filter to only records related to spending.
  Change the expense value from negative to positive for consistency.
}
\keyword{internal}
