% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarterlyIncome.R
\name{percentOverlap}
\alias{percentOverlap}
\title{percentOverlap}
\usage{
percentOverlap(interval1, interval2, units)
}
\arguments{
\item{interval1, interval2}{lubridate interval object with
specified start and end dates.}

\item{units}{character string. Units in which the results are desired.}
}
\description{
The proportion of interval 1 that overlaps the intersection
  between the two intervals.
}
