% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarterlyIncome.R
\name{quarterly_intervals}
\alias{quarterly_intervals}
\title{quarterly_intervals}
\usage{
quarterly_intervals(end, cutoff = NULL, inclPartial = TRUE)
}
\arguments{
\item{end}{date ending of last quarter}

\item{cutoff}{date cutoff to exclude later quarters}

\item{inclPartial}{logical to include partialy complete quarters
due to cutoff (default is TRUE)}
}
\value{
lubridate intervals
}
\description{
Find the four quarters for the year.
}
\examples{
quarterly_intervals(end = as.Date("2019-12-31"))
quarterly_intervals(end = as.Date("2019-12-31"),
                    cutoff = as.Date("2019-06-05"))
quarterly_intervals(end = as.Date("2019-12-31"),
                    cutoff = as.Date("2019-06-05"), inclPartial = FALSE)

}
