% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveLoad.R
\name{read_csvCreditCard}
\alias{read_csvCreditCard}
\title{read_csvCreditCard}
\usage{
read_csvCreditCard(file, company = c("Chase", "American Express"))
}
\arguments{
\item{file}{the name of the file path which the data are to be
read from.}

\item{company}{the name of company responsible for the
credit card. Choose among: {Chase, American Express}}
}
\value{
data.table of new credit card transactions
}
\description{
Load raw credit card csv data into R as a data.table
  and set variables to proper type for analysis.
}
\examples{
  \dontrun{
    read_csvCreditCard(file = "/budget/data/someChaseStatement.csv",
                       company = "Chase")
   }
}
