% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveLoad.R
\name{read_csvCreditCard_AmEx}
\alias{read_csvCreditCard_AmEx}
\title{read_csvCreditCard_AmEx}
\usage{
read_csvCreditCard_AmEx(file)
}
\arguments{
\item{file}{the name of the file path which the data are to be
read from.}
}
\description{
Read in American Express specific csv.
}
\keyword{internal}
