% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveLoad.R
\name{readCreditCard_Chase}
\alias{readCreditCard_Chase}
\title{readCreditCard_Chase}
\usage{
readCreditCard_Chase(file)
}
\arguments{
\item{file}{the name of the file path which the data are to be
read from.}
}
\description{
Read in Chase specific csv.
}
\keyword{internal}
