% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveLoad.R
\name{saveCreditCard}
\alias{saveCreditCard}
\title{saveCreditCard}
\usage{
saveCreditCard(path, creditCardHistory)
}
\arguments{
\item{path}{the path to the folder for the saved R credit card}

\item{creditCardHistory}{data.table of entire credit card history}
}
\value{
data.table of saved credit card history
}
\description{
Save the history of Credit Card data to \code{creditCardHistory.rda}.
  Overwrites previous save file.
}
\section{Warning}{

 Be sure that your previous home budget data has been properly loaded
 and appended before saving to disk as the contents will be overwritten
 with the new file.
}

\examples{
 \dontrun{
  saveCreditCard(path = "/budget/data",
                 creditCardHistory = myDataSet)
 }
}
