% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spendingCategories.R
\name{str_detect_any}
\alias{str_detect_any}
\title{str_detect_any}
\usage{
str_detect_any(string, pattern)
}
\arguments{
\item{string}{character vector; Input vector.}

\item{pattern}{character vector; substring or pattern to search for.}
}
\value{
logical vector of same length as input \code{string}.
}
\description{
Check if any pattern is a substring of the character vector.
Returns a logical vector indicating if any pattern
is contained within.
}
\seealso{
\code{\link[stringr:str_detect]{stringr::str_detect()}}
}
