% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveLoad.R
\name{read_csvCreditCard}
\alias{read_csvCreditCard}
\title{read_csvCreditCard}
\usage{
read_csvCreditCard(
  file,
  Date,
  Amount,
  Description,
  company = c("Chase", "American Express")
)
}
\arguments{
\item{file}{the name of the file path which the data are to be
read from.}

\item{Date}{name of the csv column header for the posting transaction date}

\item{Amount}{name of the csv column header for the transaction amount}

\item{Description}{name of the csv column header purchase description}

\item{company}{the name of company responsible for the
credit card. Choose among: {Chase, American Express}}
}
\value{
data.table of new credit card transactions
}
\description{
Load raw credit card csv data into R as a data.table
  and set variables to proper type for analysis.
}
\examples{
  \dontrun{
    read_csvCreditCard(file = "/budget/data/someChaseStatement.csv",
                       Date = "Posting Date",
                       Amount = "Amount",
                       Description = "Description",
                       company = "Chase")
   }
}
